Working grid-eye compression

+++ firmware/r3/grid-eye.c
@@ -42,7 +42,7 @@ void grideye_get(uint16_t *pixels)
 
 
 // Celcius * 4
-static uint16_t LUT[16] = {
+static int8_t LUT[16] = {
     -48,
     -24,
     -12,
@@ -73,15 +73,15 @@ static uint8_t reverse_LUT(int8_t raw)
 
 uint16_t grideye_pack(uint16_t *raw, uint8_t *done)
 {
-    int32_t average = 0;
+    uint32_t average = 0;
     for(uint8_t i=0; i<64; i++)
         average += raw[i];
     average /= 64;
 
     for(uint8_t i=0; i<32; i++)
     {
-        uint8_t high = reverse_LUT((int8_t)raw[2*i] - average);
-        uint8_t low = reverse_LUT((int8_t)raw[2*i+1] - average);
+        uint8_t high = reverse_LUT((int8_t)raw[2*i] - (int16_t)average);
+        uint8_t low = reverse_LUT((int8_t)raw[2*i+1] - (int16_t)average);
         done[i] = (high << 4) | low;
     }
     return average;
